pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "respectscraper"
version = "0.1.2"
description = "Ethical web scraper with robots.txt compliance, AI-powered interpretation, and comprehensive content extraction"
authors = [{name = "Zakhele Gamede", email = "info@zakhelegamede.com"}]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
keywords = [
    "webscraper",
    "scraping",
    "robots.txt",
    "pdf",
    "excel",
    "word",
    "ai",
    "llm",
    "ethical",
    "respectful"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "PyPDF2>=3.0.1",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "urllib3>=2.5.0",
    "openai>=1.66.3",
    "anthropic>=0.61.0",
    "pyrate-limiter>=3.9.0",
    "aiohttp>=3.12.15",
    "aiofiles>=24.1.0",
    "tqdm>=4.67.1",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]

[project.urls]
Homepage = "https://github.com/Zakhele-TechWannabe/respectscraper"
Repository = "https://github.com/Zakhele-TechWannabe/respectscraper.git"
Issues = "https://github.com/Zakhele-TechWannabe/respectscraper/issues"
Documentation = "https://github.com/Zakhele-TechWannabe/respectscraper/wiki"

[project.scripts]
respectscraper = "webscraper.cli:main"

[tool.setuptools.packages.find]
include = ["webscraper*"]

[tool.setuptools.package-data]
webscraper = ["*.json"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
